+++ kernel/src/lib/memory/heap/table-heap/structures.h
@@ -3,6 +3,9 @@
 //		part of the CORE kernel belonging to the H-OS project
 //
 
+//	IMPORTANT:	do NOT include in include.h
+	//	included in paging.h
+
 //	each allocator is 2mb in size (uses one sized page entry)
 	//	can allocate max 511 tables
 	//	uses first page for metadata
@@ -10,19 +13,19 @@
 #pragma once
 #include "../../../integers.h"
 
-typedef struct page_ptr {
+typedef struct table_ptr {
 	//	4 bytes in size (the smaller the better)
 	u32 index:		9;		//	table index in the allocator
 	u32 count:		9;		//	table count
 	u32 used:		1;
 	u32 _reserved:	13;
-} __attribute__((packed)) page_ptr;
+} __attribute__((packed)) table_ptr;
 
 //typedef any_page_table table_heap_t[512] __attribute__((aligned(0x200000)));		//	exactly 2mb of memory, aligned to 2mb
 typedef struct table_heap_t {
 
 	//	first 4kb is used as metadata
-	page_ptr ptr[512];
+	table_ptr ptr[512];
 	u64 bitmap[64];		//	each bit for each table used/free
 	heap_metadata meta;		//	metadata for *data