+++ kernel/src/lib/src/memory/heap/table-heap.c
@@ -78,8 +78,61 @@ void table_heap_reserve_memory() {
 	__builtin_unreachable();
 }
 
+any_page_table* table_alloc(table_allocator_t* alloc, u16 count) {
+	u64* const map = alloc->data->bitmap;
+	ssize_t index = bitmap_find_cleared(alloc->data->bitmap, 64, count);
+	if (index >= 0) {
+		//	found
+		bitmap_set(map, index, count);
+		table_ptr* ptr = &alloc->data->ptr[index];
+		ptr->used = true;
+		ptr->count = count;
+		ptr->index = index;
+		return &alloc->data->table[index];
+	}
+	return null;
+}
+
+void table_free(table_allocator_t* alloc, any_page_table* table) {
+	table = (any_page_table*)align(table, 4096);
+	const size_t index = (((size_t)table - (size_t)alloc->data) / sizeof(any_page_table)) - 1;
+	table_ptr* ptr = &alloc->data->ptr[index];
+	u16 count = ptr->count;
+	ptr->used = false;
+
+	bitmap_clear(alloc->data->bitmap, index, count);
+}
 
 
-void table_heap_debug([[maybe_unused]] table_allocator_t* alloc) {
+void table_shrink(table_allocator_t* alloc, any_page_table* table, u16 count) {
+	const size_t index = (((size_t)table - (size_t)alloc->data)/sizeof(any_page_table)) - 1;
+	table_ptr* ptr = &alloc->data->ptr[index];
+	if (count >= ptr->count) {
+		return;
+	}
+	const size_t c = ptr->count;
+	ptr->count = count;
+	bitmap_clear(alloc->data->bitmap, index + count, c - count);
+}
 
+
+void table_heap_debug(const table_allocator_t* alloc) {
+	//	prints only used blocks
+	const u32 c = output.color;
+	output.color = col.cyan;
+	print("table allocator "); printp((void*)alloc); printl(" scheme:");
+	if (alloc->data == null) {
+		printl("\tNULL");
+		output.color = c;
+		return;
+	}
+	const table_ptr* ptr = alloc->data->ptr;
+	for (size_t i = 0; i < TABLE_HEAP_TABLE_COUNT; i++) {
+		if (ptr[i].used) {
+			tab(); printu(i); print(":\tcount:\t"); printu(ptr->count); endl();
+			print("\t\t"); print("address:\t"); printp(&alloc->data->table[i]); endl();
+			i += ptr->count;
+		}
+	}
+	output.color = c;
 }